requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0
bcrypt==4.3.0
build==1.2.2.post1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
exceptiongroup==1.3.0
Faker==37.4.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==25.0
paramiko==3.5.1
pip-tools==7.4.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.4.0
python-dotenv==1.1.0
requests==2.32.4
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
sshtunnel==0.4.0
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
